ancient-scripts-datasets/cognate_pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cognate-pipeline"
version = "0.1.0"
description = "Cross-linguistic cognate detection pipeline with provenance tracking"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.12,<1",
    "pydantic>=2.5,<3",
    "pyyaml>=6,<7",
    "orjson>=3.9,<4",
    "sqlalchemy>=2.0,<3",
    "geoalchemy2>=0.14,<1",
    "psycopg[binary]>=3.1,<4",
    "alembic>=1.13,<2",
]

[project.optional-dependencies]
lingpy = ["lingpy>=2.6.12,<3"]
epitran = ["epitran>=1.24,<2"]
cldf = ["pycldf>=1.35,<2", "clldutils>=3.5,<4"]
phonemizer = ["phonemizer>=3.2,<4"]
all = [
    "cognate-pipeline[lingpy,epitran,cldf,phonemizer]",
]
dev = [
    "pytest>=8,<9",
    "pytest-cov>=5,<6",
]

[project.scripts]
cognate-pipeline = "cognate_pipeline.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cognate_pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring external services",
]
